api/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Processing and APIs
pyyaml>=6.0
requests>=2.31.0
aiofiles>=23.0
aiohttp>=3.9.0
httpx>=0.25.0

# Kubernetes
dnspython>=2.0.0

# Database
# Note: SQLite3 is built into Python, no need to install separately

pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
cryptography>=41.0.0
pyjwt>=2.8.0
click>=8.1.0
colorlog>=6.8.0

# Logging and Monitoring
python-json-logger>=2.0.0

# Plugin Dependencies - Version Control
PyGithub>=2.0
python-gitlab>=3.15.0

# Plugin Dependencies - Issue Tracking
jira>=3.10.0

# Plugin Dependencies - Remote Execution
paramiko>=3.4.0
fabric>=3.2.0

# Plugin Dependencies - HashiCorp Stack
python-hcl2>=4.3.0

# Plugin Dependencies - AI Tooling
agent-client-protocol

# MCP Server
mcp>=1.0.0
